Don't report an error for delete if the container is gone
When a container's main process exits, conmon typically calls 'podman
cleanup' to report the state change. For autoremove containers, this will
also delete the OCI runtime state. If the exit was triggered by something
like 'podman stop', this can cause a race where both 'podman delete' and
'podman stop' try to delete the runtime state. To avoid reporting an error
in this case, just exit if the runtime state is gone.