Add a simple CMake build system.
Bazel depends on OpenJDK, which doesn't work on FreeBSD/AArch64.  This
is sufficiently allow me to build on FreeBSD/AArch64.

The two dependencies are in ports, so just use find-package and rely on
the ports system to install them (which lets the port rebuild
automatically on security fixes in the dependencies).