Exit the child process immediately after failing to validate
If the container command can't be executed, the child (which is running
the jail) sends a status message to the parent, allowing it to report
the error back to Podman/Buildah etc. The child should then exit but it
was mistakenly waiting for the OCI start method to be called before
exiting. This never happened since create failed.

In the parent process, if the child sends a non-zero status indicating a
failure to validate the container command, we also need to clean up any
mounted volumes since the high-level will not call the OCI delete method
if create fails.